Refactor ag-session tests into focused modules
- Split session, orchestration, personality, and transcript notice tests into dedicated modules.
- Preserve coverage while consolidating model assertions and expanding setting-key and fingerprint checks.
- Keep production behavior unchanged while organizing related helpers, constants, and implementations.
