+++ crates/ag-session/src/orchestration_test.rs
@@ -1,0 +1,8 @@
+#[path = "orchestration_test/lifecycle_test.rs"]
+mod lifecycle;
+
+#[path = "orchestration_test/policy_test.rs"]
+mod policy;
+
+#[path = "orchestration_test/plan_test.rs"]
+mod plan;
