+++ crates/ag-session/src/question.rs
@@ -15,25 +15,5 @@ pub fn default_option_index(questions: &[QuestionItem], question_index: usize) -
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn default_option_index_selects_only_predefined_options() {
-        // Arrange
-        let questions = [
-            QuestionItem::with_options("Continue?", vec!["Yes".to_string()]),
-            QuestionItem::new("Why?"),
-        ];
-
-        // Act
-        let predefined = default_option_index(&questions, 0);
-        let free_text = default_option_index(&questions, 1);
-        let missing = default_option_index(&questions, 2);
-
-        // Assert
-        assert_eq!(predefined, Some(0));
-        assert_eq!(free_text, None);
-        assert_eq!(missing, None);
-    }
-}
+#[path = "question_test.rs"]
+mod tests;
