+++ crates/ag-session/src/service.rs
@@ -4,7 +4,8 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 
-use crate::{ReviewRequest, Session, SessionError, SessionId};
+use crate::error::SessionError;
+use crate::model::{ReviewRequest, Session, SessionId};
 
 /// Creation strategy for a new session.
 #[derive(Clone, Debug, Default, Eq, PartialEq)]
@@ -240,421 +241,5 @@ impl SessionService {
 }
 
 #[cfg(test)]
-mod tests {
-    use std::collections::VecDeque;
-    use std::sync::Mutex;
-
-    use ag_agent::{
-        AgentKind, AgentModel, AgentSelection, ReasoningLevel, ResponseStyle, SpeedMode,
-    };
-    use ag_forge::{ForgeKind, ReviewRequestState, ReviewRequestSummary};
-
-    use super::*;
-    use crate::{
-        PermissionMode, SessionMessage, SessionMessageKind, SessionRole, SessionSettings,
-        SessionStatus,
-    };
-
-    #[derive(Default)]
-    struct FakeBackend {
-        state: Mutex<FakeBackendState>,
-    }
-
-    impl FakeBackend {
-        fn from_state(state: FakeBackendState) -> Self {
-            Self {
-                state: Mutex::new(state),
-            }
-        }
-
-        fn calls(&self) -> Vec<String> {
-            self.state
-                .lock()
-                .map(|state| state.calls.clone())
-                .unwrap_or_default()
-        }
-    }
-
-    #[derive(Default)]
-    struct FakeBackendState {
-        calls: Vec<String>,
-        create_results: VecDeque<Result<SessionId, SessionError>>,
-        get_result: Option<Result<Option<Session>, SessionError>>,
-        review_result: Option<Result<ReviewRequest, SessionError>>,
-        unit_results: VecDeque<Result<(), SessionError>>,
-    }
-
-    #[async_trait]
-    impl SessionBackend for FakeBackend {
-        async fn create_session(
-            &self,
-            request: CreateSessionRequest,
-        ) -> Result<SessionId, SessionError> {
-            let mut state = self
-                .state
-                .lock()
-                .expect("fake backend state should remain available");
-            state.calls.push(format!("create:{:?}", request.mode));
-
-            state
-                .create_results
-                .pop_front()
-                .unwrap_or_else(|| Err(SessionError::Operation("missing result".to_string())))
-        }
-
-        async fn get_session(
-            &self,
-            _session_id: &SessionId,
-        ) -> Result<Option<Session>, SessionError> {
-            self.state
-                .lock()
-                .expect("fake backend state should remain available")
-                .get_result
-                .clone()
-                .unwrap_or_else(|| Err(SessionError::Operation("missing result".to_string())))
-        }
-
-        async fn send_message(
-            &self,
-            session_id: &SessionId,
-            message: String,
-        ) -> Result<(), SessionError> {
-            let mut state = self
-                .state
-                .lock()
-                .expect("fake backend state should remain available");
-            state.calls.push(format!("send:{session_id}:{message}"));
-
-            state
-                .unit_results
-                .pop_front()
-                .unwrap_or_else(|| Err(SessionError::Operation("missing result".to_string())))
-        }
-
-        async fn submit_coordinator_message(
-            &self,
-            session_id: &SessionId,
-            request: CoordinatorMessageRequest,
-        ) -> Result<(), SessionError> {
-            let mut state = self
-                .state
-                .lock()
-                .expect("fake backend state should remain available");
-            state.calls.push(format!(
-                "submit-coordinator:{session_id}:{}:{}",
-                request.operation_id, request.message
-            ));
-
-            state
-                .unit_results
-                .pop_front()
-                .unwrap_or_else(|| Err(SessionError::Operation("missing result".to_string())))
-        }
-
-        async fn answer_questions(
-            &self,
-            session_id: &SessionId,
-            request: AnswerQuestionsRequest,
-        ) -> Result<(), SessionError> {
-            let mut state = self
-                .state
-                .lock()
-                .expect("fake backend state should remain available");
-            state
-                .calls
-                .push(format!("answer:{session_id}:{}", request.answers.len()));
-
-            state
-                .unit_results
-                .pop_front()
-                .unwrap_or_else(|| Err(SessionError::Operation("missing result".to_string())))
-        }
-
-        async fn cancel_session(&self, session_id: &SessionId) -> Result<(), SessionError> {
-            let mut state = self
-                .state
-                .lock()
-                .expect("fake backend state should remain available");
-            state.calls.push(format!("cancel:{session_id}"));
-
-            state
-                .unit_results
-                .pop_front()
-                .unwrap_or_else(|| Err(SessionError::Operation("missing result".to_string())))
-        }
-
-        async fn merge_session(&self, session_id: &SessionId) -> Result<(), SessionError> {
-            let mut state = self
-                .state
-                .lock()
-                .expect("fake backend state should remain available");
-            state.calls.push(format!("merge:{session_id}"));
-
-            state
-                .unit_results
-                .pop_front()
-                .unwrap_or_else(|| Err(SessionError::Operation("missing result".to_string())))
-        }
-
-        async fn create_review_request(
-            &self,
-            session_id: &SessionId,
-        ) -> Result<ReviewRequest, SessionError> {
-            let mut state = self
-                .state
-                .lock()
-                .expect("fake backend state should remain available");
-            state.calls.push(format!("review:{session_id}"));
-
-            state
-                .review_result
-                .clone()
-                .unwrap_or_else(|| Err(SessionError::Operation("missing result".to_string())))
-        }
-    }
-
-    fn session_fixture() -> Session {
-        Session {
-            created_at: 10,
-            draft_prompt: None,
-            id: SessionId::from("session-1"),
-            messages: vec![SessionMessage::new(
-                0,
-                SessionMessageKind::UserPrompt,
-                "build it",
-            )],
-            published_upstream_ref: None,
-            questions: Vec::new(),
-            queued_messages: Vec::new(),
-            review_request: None,
-            settings: SessionSettings {
-                agent: AgentSelection::new(AgentKind::Codex, AgentModel::Gpt56Sol),
-                base_branch: "main".to_string(),
-                is_draft: false,
-                parent_session_id: None,
-                permission_mode: PermissionMode::AutoEdit,
-                personality_id: Some("reviewer".to_string()),
-                project_id: 7,
-                reasoning_level: ReasoningLevel::High,
-                response_style: ResponseStyle::Balanced,
-                role: SessionRole::Worker,
-                speed_mode: SpeedMode::Normal,
-            },
-            status: SessionStatus::Review,
-            title: Some("Build it".to_string()),
-            updated_at: 20,
-        }
-    }
-
-    fn review_request_fixture() -> ReviewRequest {
-        ReviewRequest {
-            last_refreshed_at: 30,
-            summary: ReviewRequestSummary {
-                display_id: "#42".to_string(),
-                forge_kind: ForgeKind::GitHub,
-                source_branch: "wt/session-1".to_string(),
-                state: ReviewRequestState::Open,
-                status_summary: None,
-                target_branch: "main".to_string(),
-                title: "Build it".to_string(),
-                web_url: "https://example.test/pull/42".to_string(),
-            },
-        }
-    }
-
-    #[tokio::test]
-    async fn service_delegates_create_and_get() {
-        // Arrange
-        let expected_session = session_fixture();
-        let backend = Arc::new(FakeBackend::from_state(FakeBackendState {
-            create_results: VecDeque::from([Ok(SessionId::from("session-1"))]),
-            get_result: Some(Ok(Some(expected_session.clone()))),
-            ..FakeBackendState::default()
-        }));
-        let service = SessionService::new(backend.clone());
-
-        // Act
-        let session_id = service
-            .create_session(CreateSessionRequest {
-                inherit_from_session_id: None,
-                mode: CreateSessionMode::Regular,
-                project_id: 7,
-            })
-            .await
-            .expect("session should be created");
-        let loaded_session = service
-            .get_session(&session_id)
-            .await
-            .expect("session should load");
-
-        // Assert
-        assert_eq!(loaded_session, Some(expected_session));
-        assert_eq!(backend.calls(), ["create:Regular"]);
-    }
-
-    #[tokio::test]
-    async fn service_delegates_mutating_operations_through_clones() {
-        // Arrange
-        let expected_review_request = review_request_fixture();
-        let backend = Arc::new(FakeBackend::from_state(FakeBackendState {
-            review_result: Some(Ok(expected_review_request.clone())),
-            unit_results: VecDeque::from([Ok(()), Ok(()), Ok(()), Ok(()), Ok(())]),
-            ..FakeBackendState::default()
-        }));
-        let session_id = SessionId::from("session-1");
-        let service = SessionService::new(backend.clone());
-        let cloned_service = service.clone();
-        let answers = AnswerQuestionsRequest {
-            answers: vec![QuestionAnswer {
-                answer: "main".to_string(),
-                question: "Which branch?".to_string(),
-            }],
-        };
-
-        // Act
-        service
-            .send_message(&session_id, "continue")
-            .await
-            .expect("message should be sent");
-        service
-            .submit_coordinator_message(
-                &session_id,
-                CoordinatorMessageRequest {
-                    message: "roll up".to_string(),
-                    operation_id: "rollup-7".to_string(),
-                    visibility: CoordinatorMessageVisibility::Hidden,
-                },
-            )
-            .await
-            .expect("coordinator message should be submitted");
-        cloned_service
-            .answer_questions(&session_id, answers)
-            .await
-            .expect("questions should be answered");
-        service
-            .cancel_session(&session_id)
-            .await
-            .expect("cancel should be requested");
-        cloned_service
-            .merge_session(&session_id)
-            .await
-            .expect("merge should be requested");
-        let review_request = service
-            .create_review_request(&session_id)
-            .await
-            .expect("review request should be created");
-
-        // Assert
-        assert_eq!(review_request, expected_review_request);
-        assert_eq!(
-            backend.calls(),
-            [
-                "send:session-1:continue",
-                "submit-coordinator:session-1:rollup-7:roll up",
-                "answer:session-1:1",
-                "cancel:session-1",
-                "merge:session-1",
-                "review:session-1"
-            ]
-        );
-    }
-
-    #[tokio::test]
-    async fn service_preserves_backend_errors() {
-        // Arrange
-        let expected_error = SessionError::Operation("cannot create".to_string());
-        let backend = Arc::new(FakeBackend::from_state(FakeBackendState {
-            create_results: VecDeque::from([Err(expected_error.clone())]),
-            ..FakeBackendState::default()
-        }));
-        let service = SessionService::new(backend);
-
-        // Act
-        let error = service
-            .create_session(CreateSessionRequest {
-                inherit_from_session_id: None,
-                mode: CreateSessionMode::Draft,
-                project_id: 7,
-            })
-            .await
-            .expect_err("backend error should be preserved");
-
-        // Assert
-        assert_eq!(error, expected_error);
-    }
-
-    #[tokio::test]
-    async fn fake_backend_requires_explicit_results() {
-        // Arrange
-        let backend = Arc::new(FakeBackend::default());
-        let session_id = SessionId::from("session-1");
-        let service = SessionService::new(backend);
-
-        // Act
-        let create_error = service
-            .create_session(CreateSessionRequest {
-                inherit_from_session_id: None,
-                mode: CreateSessionMode::Regular,
-                project_id: 7,
-            })
-            .await
-            .expect_err("create should require a result");
-        let get_error = service
-            .get_session(&session_id)
-            .await
-            .expect_err("get should require a result");
-        let send_error = service
-            .send_message(&session_id, "continue")
-            .await
-            .expect_err("send should require a result");
-        let coordinator_error = service
-            .submit_coordinator_message(
-                &session_id,
-                CoordinatorMessageRequest {
-                    message: "roll up".to_string(),
-                    operation_id: "rollup-1".to_string(),
-                    visibility: CoordinatorMessageVisibility::Hidden,
-                },
-            )
-            .await
-            .expect_err("coordinator submission should require a result");
-        let answer_error = service
-            .answer_questions(
-                &session_id,
-                AnswerQuestionsRequest {
-                    answers: Vec::new(),
-                },
-            )
-            .await
-            .expect_err("answers should require a result");
-        let cancel_error = service
-            .cancel_session(&session_id)
-            .await
-            .expect_err("cancel should require a result");
-        let merge_error = service
-            .merge_session(&session_id)
-            .await
-            .expect_err("merge should require a result");
-        let review_error = service
-            .create_review_request(&session_id)
-            .await
-            .expect_err("review should require a result");
-        let errors = [
-            create_error,
-            get_error,
-            send_error,
-            coordinator_error,
-            answer_error,
-            cancel_error,
-            merge_error,
-            review_error,
-        ];
-
-        // Assert
-        assert!(
-            errors
-                .into_iter()
-                .all(|error| error == SessionError::Operation("missing result".to_string()))
-        );
-    }
-}
+#[path = "service_test.rs"]
+mod tests;
