+++ crates/ag-session/src/service_test.rs
@@ -1,0 +1,8 @@
+#[path = "service_test/delegation_test.rs"]
+mod delegation;
+
+#[path = "service_test/error_test.rs"]
+mod error;
+
+#[path = "service_test/support_test.rs"]
+mod support;
